Demo/requirements.txt
analytics-python==1.4.0
backoff==1.10.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
clip @ git+https://github.com/openai/CLIP.git@573315e83f07b53a61ff5098757e8fc885f1703e
cryptography==36.0.1
cycler==0.11.0
faiss-cpu==1.7.2
ffmpy==0.3.0
filelock==3.4.2
Flask==2.0.2
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
fonttools==4.28.5
ftfy==6.0.3
gitdb==4.0.9
GitPython==3.1.26
gradio==2.7.0
idna==3.3
imageio==2.13.5
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
markdown2==2.4.2
MarkupSafe==2.0.1
matplotlib==3.5.1
monotonic==1.6
networkx==2.6.3
numpy==1.22.1
packaging==21.3
pandas==1.3.5
paramiko==2.9.2
Pillow==9.0.0
pycparser==2.21
pycryptodome==3.12.0
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.6
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
regex==2021.11.10
requests==2.27.1
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
tifffile==2021.11.2
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
gdown